libm: gtests: Fixed Ulps getting rounded off to int
    Due to this, ulp print is always rounding off to the
	nearest interger and not printing actual ulp

+++ gtests/gtest_main.cc
@@ -162,7 +162,7 @@ TEST_P(AccuTestFixtureDouble, ACCURACY_SCALAR_DOUBLE) {
       ip[1] = inpbuff1[i];
 
     double exptd = getExpected(ip);
-    double ulp = getUlp(aop[0], exptd);
+    double ulp = getUlp(aop[0], (long double)exptd);
     if(!update_ulp(ulp, max_ulp_err, inData->ulp_threshold)) {
       nfail++;
     }
@@ -190,20 +190,20 @@ TEST_P(AccuTestFixtureDouble, ACCURACY_VECTOR_2DOUBLES) {
   data.ip  = (void *)inpbuff;
   data.op  = (void *)aop;
   double ip[2];
-  
+
   if(nargs == 2)
     data.ip1 = (void *)inpbuff1;
-  
+
   for (uint32_t i = 0; i < count; i += 2) {
     test_v2d(&data, i);
 
     for (uint32_t j = 0; j < 2; j++) {
       ip[0] = inpbuff[i + j];
-      if(nargs == 2)    
+      if(nargs == 2) 
         ip[1] = inpbuff1[i + j];
 
       double exptd = getExpected(ip);
-      double ulp = getUlp(aop[j], exptd);
+      double ulp = getUlp(aop[j], (long double)exptd);
       if(!update_ulp(ulp, max_ulp_err, inData->ulp_threshold)) {
         nfail++;
       }
@@ -245,7 +245,7 @@ TEST_P(AccuTestFixtureDouble, ACCURACY_VECTOR_4DOUBLES) {
         ip[1] = inpbuff1[i + j];
 
       double exptd = getExpected(ip);
-      double ulp = getUlp(aop[j], exptd);
+      double ulp = getUlp(aop[j], (long double)exptd);
       if(!update_ulp(ulp, max_ulp_err, inData->ulp_threshold)) {
         nfail++;
       }
